ARRISXB3-9702 ARRISXB3-9748 RDKB-26335 rbus crash only in rbus mode
Reason for change:
report rbus crash and reboot the device
only in rbus mode and on actual crash only
Test Procedure: Unit Test
Risks: Low

(cherry picked from commit 1153555d314f2101dfe985f603bb794ce810710f)

+++ scripts/rbus_termination_handler.sh
@@ -19,4 +19,9 @@
 #######################################################################################
 
 source /etc/device.properties
-rpcclient $ARM_ARPING_IP "syscfg set X_RDKCENTRAL-COM_LastRebootReason Rbus_crash && syscfg set X_RDKCENTRAL-COM_LastRebootCounter 1 && syscfg commit && sync && reboot"
+
+sleep 60
+#set rbus crash only in rbus mode
+if [ -e /nvram/rbus ]; then
+    rpcclient $ARM_ARPING_IP "syscfg set X_RDKCENTRAL-COM_LastRebootReason Rbus_crash && syscfg set X_RDKCENTRAL-COM_LastRebootCounter 1 && syscfg commit && sync && reboot"
+fi
